Enum tor_chanmgr::Error

#[non_exhaustive]
pub enum Error {

    UnusableTarget(Bug),
    PendingFailed {
        peer: LoggedChanTarget,
    },
    ChanTimeout {
        peer: LoggedChanTarget,
    },
    Proto {
        source: Error,
        peer: LoggedChanTarget,
        clock_skew: Option<ClockSkew>,
    },
    Io {
        peer: Option<BoxChanSensitive<BridgeAddr>>,
        action: &'static str,
        source: Arc<Error>,
    },
    ChannelBuild {
        addresses: Vec<(ChanSensitive<SocketAddr>, Arc<Error>)>,
    },
    Spawn {
        spawning: &'static str,
        cause: Arc<SpawnError>,
    },
    MissingId,
    IdentityConflict,
    NoSuchTransport(TransportId),
    RequestCancelled,
    Proxy(ProxyError),
    Pt(Arc<dyn AbstractPtError>),
    Internal(Bug),

}

An error returned by a channel manager.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

UnusableTarget(Bug)

A ChanTarget was given for which no channel could be built.

PendingFailed

Fields

peer: LoggedChanTarget

Who we were talking to

We were waiting on a pending channel, but it didnโt succeed.

ChanTimeout

Fields

peer: LoggedChanTarget

Who we were trying to talk to

It took too long for us to establish this connection.

Proto

Fields

source: Error

The underlying error

peer: LoggedChanTarget

Who we were trying to talk to

clock_skew: Option<ClockSkew>

An authenticated ClockSkew (if available) that we received from the peer.

A protocol error while making a channel

Io

Fields

peer: Option<BoxChanSensitive<BridgeAddr>>

Who we were talking to

action: &'static str

What we were doing

source: Arc<Error>

What happened. Might be some TLS library error wrapped up in io::Error

Network IO error or TLS error

ChannelBuild

Fields

addresses: Vec<(ChanSensitive<SocketAddr>, Arc<Error>)>

The list of addresses we tried to connect to, coupled with the error we encountered connecting to each one.

Failed to build a channel, after trying multiple addresses.

Spawn

Fields

spawning: &'static str

What we were trying to spawn.

cause: Arc<SpawnError>

What happened when we tried to spawn it.

Unable to spawn task

MissingId

A relay did not have the set of identity keys that we expected.

(Currently, tor-chanmgr only works on relays that have at least one recognized identity key.)

IdentityConflict

A successful relay channel had one of the identity keys we wanted, but not the other(s).

This means that (assuming the relay is well behaved), we will not find the ID combination we want.

NoSuchTransport(TransportId)

Tried to connect via a transport that we donโt support.

RequestCancelled

An attempt to open a channel failed because it was cancelled or superseded by another request or configuration change.

Proxy(ProxyError)

We tried to create a channel through a proxy, and encountered an error.

Pt(Arc<dyn AbstractPtError>)

An error occurred in a pluggable transport manager.

We canโt know the type, because any pluggable transport manager implementing AbstractPtMgr can be used. However, if youโre using Arti in the standard configuration, this will be tor-ptmgrโs PtError.

Internal(Bug)

An internal error of some kind that should never occur.

Implementations

impl Error

pub fn clock_skew(&self) -> Option<ClockSkew>

Return the clock skew information from this error (or from an internal error).

Only returns the clock skew information if it is authenticated.